backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
firebase-admin==7.1.0
google-cloud-firestore==2.22.0
